# internal/api/server.generated.go
//go:build go1.22

// Package api provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.8.0 DO NOT EDIT.
package api

import (
	"encoding/json"
	"errors"
	"fmt"
	"net/http"
	"time"

	"github.com/oapi-codegen/runtime"
	openapi_types "github.com/oapi-codegen/runtime/types"
)

// Defines values for HistoryEntryField.
const (
	HistoryEntryFieldAuthors         HistoryEntryField = "authors"
	HistoryEntryFieldDescription     HistoryEntryField = "description"
	HistoryEntryFieldPublicationDate HistoryEntryField = "publication_date"
	HistoryEntryFieldTitle           HistoryEntryField = "title"
)

// Valid indicates whether the value is a known member of the HistoryEntryField enum.
func (e HistoryEntryField) Valid() bool {
	switch e {
	case HistoryEntryFieldAuthors:
		return true
	case HistoryEntryFieldDescription:
		return true
	case HistoryEntryFieldPublicationDate:
		return true
	case HistoryEntryFieldTitle:
		return true
	default:
		return false
	}
}

// Defines values for HistoryEntryKind.
const (
	HistoryEntryKindCreated HistoryEntryKind = "created"
	HistoryEntryKindUpdated HistoryEntryKind = "updated"
)

// Valid indicates whether the value is a known member of the HistoryEntryKind enum.
func (e HistoryEntryKind) Valid() bool {
	switch e {
	case HistoryEntryKindCreated:
		return true
	case HistoryEntryKindUpdated:
		return true
	default:
		return false
	}
}

// Defines values for GetBookHistoryParamsKind.
const (
	GetBookHistoryParamsKindCreated GetBookHistoryParamsKind = "created"
	GetBookHistoryParamsKindUpdated GetBookHistoryParamsKind = "updated"
)

// Valid indicates whether the value is a known member of the GetBookHistoryParamsKind enum.
func (e GetBookHistoryParamsKind) Valid() bool {
	switch e {
	case GetBookHistoryParamsKindCreated:
		return true
	case GetBookHistoryParamsKindUpdated:
		return true
	default:
		return false
	}
}

// Defines values for GetBookHistoryParamsField.
const (
	GetBookHistoryParamsFieldAuthors         GetBookHistoryParamsField = "authors"
	GetBookHistoryParamsFieldDescription     GetBookHistoryParamsField = "description"
	GetBookHistoryParamsFieldPublicationDate GetBookHistoryParamsField = "publication_date"
	GetBookHistoryParamsFieldTitle           GetBookHistoryParamsField = "title"
)

// Valid indicates whether the value is a known member of the GetBookHistoryParamsField enum.
func (e GetBookHistoryParamsField) Valid() bool {
	switch e {
	case GetBookHistoryParamsFieldAuthors:
		return true
	case GetBookHistoryParamsFieldDescription:
		return true
	case GetBookHistoryParamsFieldPublicationDate:
		return true
	case GetBookHistoryParamsFieldTitle:
		return true
	default:
		return false
	}
}

// Defines values for GetBookHistoryParamsOrder.
const (
	Asc  GetBookHistoryParamsOrder = "asc"
	Desc GetBookHistoryParamsOrder = "desc"
)

// Valid indicates whether the value is a known member of the GetBookHistoryParamsOrder enum.
func (e GetBookHistoryParamsOrder) Valid() bool {
	switch e {
	case Asc:
		return true
	case Desc:
		return true
	default:
		return false
	}
}

// Authors defines model for Authors.
type Authors = []string

// Book defines model for Book.
type Book struct {
	Authors         Authors         `json:"authors"`
	CreatedAt       time.Time       `json:"created_at"`
	Description     BookDescription `json:"description"`
	Id              BookID          `json:"id"`
	PublicationDate PublicationDate `json:"publication_date"`
	Title           BookTitle       `json:"title"`
	UpdatedAt       time.Time       `json:"updated_at"`
	Version         BookVersion     `json:"version"`
}

// BookDescription defines model for BookDescription.
type BookDescription = string

// BookFields defines model for BookFields.
type BookFields struct {
	Authors         Authors         `json:"authors"`
	Description     BookDescription `json:"description"`
	PublicationDate PublicationDate `json:"publication_date"`
	Title           BookTitle       `json:"title"`
}

// BookID defines model for BookID.
type BookID = openapi_types.UUID

// BookTitle defines model for BookTitle.
type BookTitle = string

// BookVersion defines model for BookVersion.
type BookVersion = int64

// ErrorResponse defines model for ErrorResponse.
type ErrorResponse struct {
	Error struct {
		Code    string `json:"code"`
		Message string `json:"message"`
	} `json:"error"`
}

// HistoryEntry defines model for HistoryEntry.
type HistoryEntry struct {
	BookId BookID `json:"book_id"`

	// Description Human-readable text frozen at write time
	Description string            `json:"description"`
	Field       HistoryEntryField `json:"field"`
	Id          int64             `json:"id"`
	Kind        HistoryEntryKind  `json:"kind"`

	// New New JSON value
	New        HistoryValue `json:"new"`
	OccurredAt time.Time    `json:"occurred_at"`

	// Old Previous JSON value, null for initial creation
	Old *HistoryValue `json:"old"`
}

// HistoryEntryField defines model for HistoryEntry.Field.
type HistoryEntryField string

// HistoryEntryKind defines model for HistoryEntry.Kind.
type HistoryEntryKind string

// HistoryPage defines model for HistoryPage.
type HistoryPage struct {
	Items  []HistoryEntry `json:"items"`
	Limit  int            `json:"limit"`
	Offset int            `json:"offset"`
	Total  int            `json:"total"`
}

// HistoryValue A book field value
type HistoryValue struct {
	union json.RawMessage
}

// PublicationDate defines model for PublicationDate.
type PublicationDate = openapi_types.Date

// UpdateBookRequest defines model for UpdateBookRequest.
type UpdateBookRequest struct {
	// Authors Complete ordered author list
	Authors         Authors         `json:"authors"`
	Description     BookDescription `json:"description"`
	PublicationDate PublicationDate `json:"publication_date"`
	Title           BookTitle       `json:"title"`

	// Version Last version retrieved by the client
	Version BookVersion `json:"version"`
}

// BadRequest defines model for BadRequest.
type BadRequest = ErrorResponse

// Conflict defines model for Conflict.
type Conflict = ErrorResponse

// InternalError defines model for InternalError.
type InternalError = ErrorResponse

// NotFound defines model for NotFound.
type NotFound = ErrorResponse

// TooLarge defines model for TooLarge.
type TooLarge = ErrorResponse

// GetBookHistoryParams defines parameters for GetBookHistory.
type GetBookHistoryParams struct {
	Limit  *int                       `form:"limit,omitempty" json:"limit,omitempty"`
	Offset *int                       `form:"offset,omitempty" json:"offset,omitempty"`
	Kind   *GetBookHistoryParamsKind  `form:"kind,omitempty" json:"kind,omitempty"`
	Field  *GetBookHistoryParamsField `form:"field,omitempty" json:"field,omitempty"`
	From   *time.Time                 `form:"from,omitempty" json:"from,omitempty"`
	To     *time.Time                 `form:"to,omitempty" json:"to,omitempty"`
	Order  *GetBookHistoryParamsOrder `form:"order,omitempty" json:"order,omitempty"`
}

// GetBookHistoryParamsKind defines parameters for GetBookHistory.
type GetBookHistoryParamsKind string

// GetBookHistoryParamsField defines parameters for GetBookHistory.
type GetBookHistoryParamsField string

// GetBookHistoryParamsOrder defines parameters for GetBookHistory.
type GetBookHistoryParamsOrder string

// CreateBookJSONRequestBody defines body for CreateBook for application/json ContentType.
type CreateBookJSONRequestBody = BookFields

// UpdateBookJSONRequestBody defines body for UpdateBook for application/json ContentType.
type UpdateBookJSONRequestBody = UpdateBookRequest

// AsBookTitle returns the union data inside the HistoryValue as a BookTitle
func (t HistoryValue) AsBookTitle() (BookTitle, error) {
	var body BookTitle
	err := json.Unmarshal(t.union, &body)
	return body, err
}

// FromBookTitle overwrites any union data inside the HistoryValue as the provided BookTitle
func (t *HistoryValue) FromBookTitle(v BookTitle) error {
	b, err := json.Marshal(v)
	t.union = b
	return err
}

// MergeBookTitle performs a merge with any union data inside the HistoryValue, using the provided BookTitle
func (t *HistoryValue) MergeBookTitle(v BookTitle) error {
	b, err := json.Marshal(v)
	if err != nil {
		return err
	}

	merged, err := runtime.JSONMerge(t.union, b)
	t.union = merged
	return err
}

// AsBookDescription returns the union data inside the HistoryValue as a BookDescription
func (t HistoryValue) AsBookDescription() (BookDescription, error) {
	var body BookDescription
	err := json.Unmarshal(t.union, &body)
	return body, err
}

// FromBookDescription overwrites any union data inside the HistoryValue as the provided BookDescription
func (t *HistoryValue) FromBookDescription(v BookDescription) error {
	b, err := json.Marshal(v)
	t.union = b
	return err
}

// MergeBookDescription performs a merge with any union data inside the HistoryValue, using the provided BookDescription
func (t *HistoryValue) MergeBookDescription(v BookDescription) error {
	b, err := json.Marshal(v)
	if err != nil {
		return err
	}

	merged, err := runtime.JSONMerge(t.union, b)
	t.union = merged
	return err
}

// AsPublicationDate returns the union data inside the HistoryValue as a PublicationDate
func (t HistoryValue) AsPublicationDate() (PublicationDate, error) {
	var body PublicationDate
	err := json.Unmarshal(t.union, &body)
	return body, err
}

// FromPublicationDate overwrites any union data inside the HistoryValue as the provided PublicationDate
func (t *HistoryValue) FromPublicationDate(v PublicationDate) error {
	b, err := json.Marshal(v)
	t.union = b
	return err
}

// MergePublicationDate performs a merge with any union data inside the HistoryValue, using the provided PublicationDate
func (t *HistoryValue) MergePublicationDate(v PublicationDate) error {
	b, err := json.Marshal(v)
	if err != nil {
		return err
	}

	merged, err := runtime.JSONMerge(t.union, b)
	t.union = merged
	return err
}

// AsAuthors returns the union data inside the HistoryValue as a Authors
func (t HistoryValue) AsAuthors() (Authors, error) {
	var body Authors
	err := json.Unmarshal(t.union, &body)
	return body, err
}

// FromAuthors overwrites any union data inside the HistoryValue as the provided Authors
func (t *HistoryValue) FromAuthors(v Authors) error {
	b, err := json.Marshal(v)
	t.union = b
	return err
}

// MergeAuthors performs a merge with any union data inside the HistoryValue, using the provided Authors
func (t *HistoryValue) MergeAuthors(v Authors) error {
	b, err := json.Marshal(v)
	if err != nil {
		return err
	}

	merged, err := runtime.JSONMerge(t.union, b)
	t.union = merged
	return err
}

func (t HistoryValue) MarshalJSON() ([]byte, error) {
	b, err := t.union.MarshalJSON()
	return b, err
}

func (t *HistoryValue) UnmarshalJSON(b []byte) error {
	err := t.union.UnmarshalJSON(b)
	return err
}

// ServerInterface represents all server handlers.
type ServerInterface interface {
	// CreateBook Create a new book
	// (POST /api/books)
	CreateBook(w http.ResponseWriter, r *http.Request)
	// GetBook Retrieve a book
	// (GET /api/books/{id})
	GetBook(w http.ResponseWriter, r *http.Request, id BookID)
	// UpdateBook Update an existing book
	// (PUT /api/books/{id})
	UpdateBook(w http.ResponseWriter, r *http.Request, id BookID)
	// GetBookHistory Get the history of changes made to the book
	// (GET /api/books/{id}/history)
	GetBookHistory(w http.ResponseWriter, r *http.Request, id BookID, params GetBookHistoryParams)
	// Health Check service health
	// (GET /api/healthz)
	Health(w http.ResponseWriter, r *http.Request)
	// GetDocs Redirect to the interactive API UI
	// (GET /docs)
	GetDocs(w http.ResponseWriter, r *http.Request)
	// GetOpenAPISpec Retrieve the OpenAPI spec
	// (GET /docs/openapi.yaml)
	GetOpenAPISpec(w http.ResponseWriter, r *http.Request)
}

// ServerInterfaceWrapper converts contexts to parameters.
type ServerInterfaceWrapper struct {
	Handler            ServerInterface
	HandlerMiddlewares []MiddlewareFunc
	ErrorHandlerFunc   func(w http.ResponseWriter, r *http.Request, err error)
}

type MiddlewareFunc func(http.Handler) http.Handler

// CreateBook operation middleware
func (siw *ServerInterfaceWrapper) CreateBook(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.CreateBook(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// GetBook operation middleware
func (siw *ServerInterfaceWrapper) GetBook(w http.ResponseWriter, r *http.Request) {

	var err error
	_ = err

	// ------------- Path parameter "id" -------------
	var id BookID

	err = runtime.BindStyledParameterWithOptions("simple", "id", r.PathValue("id"), &id, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true, Type: "string", Format: "uuid", ValueIsUnescaped: true})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "id", Err: err})
		return
	}

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.GetBook(w, r, id)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// UpdateBook operation middleware
func (siw *ServerInterfaceWrapper) UpdateBook(w http.ResponseWriter, r *http.Request) {

	var err error
	_ = err

	// ------------- Path parameter "id" -------------
	var id BookID

	err = runtime.BindStyledParameterWithOptions("simple", "id", r.PathValue("id"), &id, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true, Type: "string", Format: "uuid", ValueIsUnescaped: true})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "id", Err: err})
		return
	}

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.UpdateBook(w, r, id)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// GetBookHistory operation middleware
func (siw *ServerInterfaceWrapper) GetBookHistory(w http.ResponseWriter, r *http.Request) {

	var err error
	_ = err

	// ------------- Path parameter "id" -------------
	var id BookID

	err = runtime.BindStyledParameterWithOptions("simple", "id", r.PathValue("id"), &id, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true, Type: "string", Format: "uuid", ValueIsUnescaped: true})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "id", Err: err})
		return
	}

	// Parameter object where we will unmarshal all parameters from the context
	var params GetBookHistoryParams

	// ------------- Optional query parameter "limit" -------------

	err = runtime.BindQueryParameterWithOptions("form", true, false, "limit", r.URL.Query(), &params.Limit, runtime.BindQueryParameterOptions{Type: "integer", Format: ""})
	if err != nil {
		var requiredError *runtime.RequiredParameterError
		if errors.As(err, &requiredError) {
			siw.ErrorHandlerFunc(w, r, &RequiredParamError{ParamName: "limit"})
		} else {
			siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "limit", Err: err})
		}
		return
	}

	// ------------- Optional query parameter "offset" -------------

	err = runtime.BindQueryParameterWithOptions("form", true, false, "offset", r.URL.Query(), &params.Offset, runtime.BindQueryParameterOptions{Type: "integer", Format: ""})
	if err != nil {
		var requiredError *runtime.RequiredParameterError
		if errors.As(err, &requiredError) {
			siw.ErrorHandlerFunc(w, r, &RequiredParamError{ParamName: "offset"})
		} else {
			siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "offset", Err: err})
		}
		return
	}

	// ------------- Optional query parameter "kind" -------------

	err = runtime.BindQueryParameterWithOptions("form", true, false, "kind", r.URL.Query(), &params.Kind, runtime.BindQueryParameterOptions{Type: "string", Format: ""})
	if err != nil {
		var requiredError *runtime.RequiredParameterError
		if errors.As(err, &requiredError) {
			siw.ErrorHandlerFunc(w, r, &RequiredParamError{ParamName: "kind"})
		} else {
			siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "kind", Err: err})
		}
		return
	}

	// ------------- Optional query parameter "field" -------------

	err = runtime.BindQueryParameterWithOptions("form", true, false, "field", r.URL.Query(), &params.Field, runtime.BindQueryParameterOptions{Type: "string", Format: ""})
	if err != nil {
		var requiredError *runtime.RequiredParameterError
		if errors.As(err, &requiredError) {
			siw.ErrorHandlerFunc(w, r, &RequiredParamError{ParamName: "field"})
		} else {
			siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "field", Err: err})
		}
		return
	}

	// ------------- Optional query parameter "from" -------------

	err = runtime.BindQueryParameterWithOptions("form", true, false, "from", r.URL.Query(), &params.From, runtime.BindQueryParameterOptions{Type: "string", Format: "date-time"})
	if err != nil {
		var requiredError *runtime.RequiredParameterError
		if errors.As(err, &requiredError) {
			siw.ErrorHandlerFunc(w, r, &RequiredParamError{ParamName: "from"})
		} else {
			siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "from", Err: err})
		}
		return
	}

	// ------------- Optional query parameter "to" -------------

	err = runtime.BindQueryParameterWithOptions("form", true, false, "to", r.URL.Query(), &params.To, runtime.BindQueryParameterOptions{Type: "string", Format: "date-time"})
	if err != nil {
		var requiredError *runtime.RequiredParameterError
		if errors.As(err, &requiredError) {
			siw.ErrorHandlerFunc(w, r, &RequiredParamError{ParamName: "to"})
		} else {
			siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "to", Err: err})
		}
		return
	}

	// ------------- Optional query parameter "order" -------------

	err = runtime.BindQueryParameterWithOptions("form", true, false, "order", r.URL.Query(), &params.Order, runtime.BindQueryParameterOptions{Type: "string", Format: ""})
	if err != nil {
		var requiredError *runtime.RequiredParameterError
		if errors.As(err, &requiredError) {
			siw.ErrorHandlerFunc(w, r, &RequiredParamError{ParamName: "order"})
		} else {
			siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "order", Err: err})
		}
		return
	}

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.GetBookHistory(w, r, id, params)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// Health operation middleware
func (siw *ServerInterfaceWrapper) Health(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.Health(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// GetDocs operation middleware
func (siw *ServerInterfaceWrapper) GetDocs(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.GetDocs(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// GetOpenAPISpec operation middleware
func (siw *ServerInterfaceWrapper) GetOpenAPISpec(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.GetOpenAPISpec(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

type UnescapedCookieParamError struct {
	ParamName string
	Err       error
}

func (e *UnescapedCookieParamError) Error() string {
	return fmt.Sprintf("error unescaping cookie parameter '%s'", e.ParamName)
}

func (e *UnescapedCookieParamError) Unwrap() error {
	return e.Err
}

type UnmarshalingParamError struct {
	ParamName string
	Err       error
}

func (e *UnmarshalingParamError) Error() string {
	return fmt.Sprintf("Error unmarshaling parameter %s as JSON: %s", e.ParamName, e.Err.Error())
}

func (e *UnmarshalingParamError) Unwrap() error {
	return e.Err
}

type RequiredParamError struct {
	ParamName string
}

func (e *RequiredParamError) Error() string {
	return fmt.Sprintf("Query argument %s is required, but not found", e.ParamName)
}

type RequiredHeaderError struct {
	ParamName string
	Err       error
}

func (e *RequiredHeaderError) Error() string {
	return fmt.Sprintf("Header parameter %s is required, but not found", e.ParamName)
}

func (e *RequiredHeaderError) Unwrap() error {
	return e.Err
}

type InvalidParamFormatError struct {
	ParamName string
	Err       error
}

func (e *InvalidParamFormatError) Error() string {
	return fmt.Sprintf("Invalid format for parameter %s: %s", e.ParamName, e.Err.Error())
}

func (e *InvalidParamFormatError) Unwrap() error {
	return e.Err
}

type TooManyValuesForParamError struct {
	ParamName string
	Count     int
}

func (e *TooManyValuesForParamError) Error() string {
	return fmt.Sprintf("Expected one value for %s, got %d", e.ParamName, e.Count)
}

// Handler creates http.Handler with routing matching OpenAPI spec.
func Handler(si ServerInterface) http.Handler {
	return HandlerWithOptions(si, StdHTTPServerOptions{})
}

// ServeMux is an abstraction of [http.ServeMux].
type ServeMux interface {
	HandleFunc(pattern string, handler func(http.ResponseWriter, *http.Request))
	http.Handler
}

type StdHTTPServerOptions struct {
	BaseURL          string
	BaseRouter       ServeMux
	Middlewares      []MiddlewareFunc
	ErrorHandlerFunc func(w http.ResponseWriter, r *http.Request, err error)
}

// HandlerFromMux creates http.Handler with routing matching OpenAPI spec based on the provided mux.
func HandlerFromMux(si ServerInterface, m ServeMux) http.Handler {
	return HandlerWithOptions(si, StdHTTPServerOptions{
		BaseRouter: m,
	})
}

func HandlerFromMuxWithBaseURL(si ServerInterface, m ServeMux, baseURL string) http.Handler {
	return HandlerWithOptions(si, StdHTTPServerOptions{
		BaseURL:    baseURL,
		BaseRouter: m,
	})
}

// HandlerWithOptions creates http.Handler with additional options
func HandlerWithOptions(si ServerInterface, options StdHTTPServerOptions) http.Handler {
	m := options.BaseRouter

	if m == nil {
		m = http.NewServeMux()
	}
	if options.ErrorHandlerFunc == nil {
		options.ErrorHandlerFunc = func(w http.ResponseWriter, r *http.Request, err error) {
			http.Error(w, err.Error(), http.StatusBadRequest)
		}
	}

	wrapper := ServerInterfaceWrapper{
		Handler:            si,
		HandlerMiddlewares: options.Middlewares,
		ErrorHandlerFunc:   options.ErrorHandlerFunc,
	}

	m.HandleFunc(http.MethodPost+" "+options.BaseURL+"/api/books", wrapper.CreateBook)
	m.HandleFunc(http.MethodGet+" "+options.BaseURL+"/api/books/{id}", wrapper.GetBook)
	m.HandleFunc(http.MethodPut+" "+options.BaseURL+"/api/books/{id}", wrapper.UpdateBook)
	m.HandleFunc(http.MethodGet+" "+options.BaseURL+"/api/books/{id}/history", wrapper.GetBookHistory)
	m.HandleFunc(http.MethodGet+" "+options.BaseURL+"/api/healthz", wrapper.Health)
	m.HandleFunc(http.MethodGet+" "+options.BaseURL+"/docs", wrapper.GetDocs)
	m.HandleFunc(http.MethodGet+" "+options.BaseURL+"/docs/openapi.yaml", wrapper.GetOpenAPISpec)

	return m
}